@@ -1,42 +1,37 @@
-# Settings specified here will take precedence over those in config/environment.rb
+WardAreaBook::Application.configure do
+  # Settings specified here will take precedence over those in config/application.rb
 
-# In the development environment your application's code is reloaded on
-# every request.  This slows down response time but is perfect for development
-# since you don't have to restart the webserver when you make code changes.
-config.cache_classes = false
+  # In the development environment your application's code is reloaded on
+  # every request. This slows down response time but is perfect for development
+  # since you don't have to restart the web server when you make code changes.
+  config.cache_classes = false
 
-# Log error messages when you accidentally call methods on nil.
-config.whiny_nils = true
+  # Log error messages when you accidentally call methods on nil.
+  config.whiny_nils = true
 
-# Show full error reports and disable caching
-config.action_controller.consider_all_requests_local = true
-config.action_view.debug_rjs                         = true
-config.action_controller.perform_caching             = false
+  # Show full error reports and disable caching
+  config.consider_all_requests_local       = true
+  config.action_controller.perform_caching = false
 
-# Mail
-# config/environments/development.rb
-config.action_mailer.raise_delivery_errors = true
+  # Don't care if the mailer can't send
+  config.action_mailer.raise_delivery_errors = false
 
-# set delivery method to :smtp, :sendmail or :test
-config.action_mailer.delivery_method = :smtp
+  # Print deprecation notices to the Rails logger
+  config.active_support.deprecation = :log
 
-=begin
-config.action_mailer.smtp_settings = {
-  :address => "xxx",
-  :authentication => :login,
-  :user_name => "xxx",
-  :password => "xxx"
-}
-=end
+  # Only use best-standards-support built into browsers
+  config.action_dispatch.best_standards_support = :builtin
 
-# these options are only needed if you choose smtp delivery
-config.action_mailer.smtp_settings = {
-  :address => "smtp.gmail.com", 
-  :port => 587, 
-  :authentication => :plain, 
-  :enable_starttls_auto => true, 
-  :user_name => "kinateder@gmail.com", 
-  :password => "Reno55Kinatron!@#" 
-}
+  # Raise exception on mass assignment protection for Active Record models
+  config.active_record.mass_assignment_sanitizer = :strict
 
-ActionController::Base.cache_store = :file_store, "#{RAILS_ROOT}/public/cache"
+  # Log the query plan for queries taking more than this (works
+  # with SQLite, MySQL, and PostgreSQL)
+  config.active_record.auto_explain_threshold_in_seconds = 0.5
+
+  # Do not compress assets
+  config.assets.compress = false
+
+  # Expands the lines which load the assets
+  config.assets.debug = true
+end
